feat: creat model and migration sales

+++ back-end/src/database/models/product.js
@@ -2,7 +2,7 @@ module.exports = (sequelize, DataTypes) => {
   const Product = sequelize.define('Product', {
    id: { type: DataTypes.INTEGER, autoIncrement: true, primaryKey: true },
    name: DataTypes.STRING(100),
-   price: DataTypes.DOUBLE(4,2),
+   price: DataTypes.DECIMAL(4,2),
    url_image: DataTypes.STRING(200),
   },
   { timestamps: false,
